refactor marker into a new module
